fix(module:grid): fix grid responsive (#2915)
close #2908
