+++ components/form/nz-form-label.component.ts
@@ -1,15 +1,16 @@
 import {
+  AfterViewInit,
   ChangeDetectionStrategy,
   Component,
   ElementRef,
   Host,
-  Input,
+  Input, OnDestroy,
   Optional, Renderer2,
   ViewEncapsulation
 } from '@angular/core';
 import { NzUpdateHostClassService } from '../core/services/update-host-class.service';
 import { InputBoolean } from '../core/util/convert';
-import { NzColComponent } from '../grid/nz-col.component';
+import { NzColDirective } from '../grid/nz-col.directive';
 import { NzRowDirective } from '../grid/nz-row.directive';
 import { NzFormItemComponent } from './nz-form-item.component';
 
@@ -21,12 +22,20 @@ import { NzFormItemComponent } from './nz-form-item.component';
   changeDetection    : ChangeDetectionStrategy.OnPush,
   templateUrl        : './nz-form-label.component.html'
 })
-export class NzFormLabelComponent extends NzColComponent {
+export class NzFormLabelComponent extends NzColDirective implements OnDestroy, AfterViewInit {
   @Input() nzFor: string;
   @Input() @InputBoolean() nzRequired = false;
 
   constructor(nzUpdateHostClassService: NzUpdateHostClassService, elementRef: ElementRef, @Optional() @Host() nzFormItemComponent: NzFormItemComponent, @Optional() @Host() nzRowDirective: NzRowDirective, renderer: Renderer2) {
-    super(nzUpdateHostClassService, elementRef, nzFormItemComponent, nzRowDirective);
+    super(nzUpdateHostClassService, elementRef, nzFormItemComponent || nzRowDirective, renderer);
     renderer.addClass(elementRef.nativeElement, 'ant-form-item-label');
+  }
+
+  ngOnDestroy(): void {
+    super.ngOnDestroy();
+  }
+
+  ngAfterViewInit(): void {
+    super.ngAfterViewInit();
   }
 }
