--- components/grid/nz-col.component.html
@@ -1,1 +1,0 @@
-<ng-content></ng-content>
