--- components/grid/nz-row.component.html
@@ -1,1 +1,0 @@
-<ng-content></ng-content>
