+++ components/grid/public-api.ts
@@ -1,5 +1,3 @@
 export * from './nz-row.directive';
-export * from './nz-row.component';
-export * from './nz-col.component';
 export * from './nz-col.directive';
 export * from './nz-grid.module';
